first

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[10];
+        long[] sales = new long[10];
         for (int i = 0; i < sales.length; i++) {
-            sales[i] = (int) ((i + 1) * Math.random()*100);
+            sales[i] = (long) ((i + 1) * Math.random()*100);
         }
         SalesManager salesManager = new SalesManager(sales);
         System.out.println("Лучшая продажа " + salesManager.max());
